+++ sgkim94/src/main/java/com/example/sgkim94/item05/NotificationFactory.java
@@ -1,4 +1,4 @@
-package com.example.sgkim94.item04;
+package com.example.sgkim94.item05;
 
 public class NotificationFactory {
     public Notification getNotification(String type) {
