+++ sgkim94/src/main/java/com/example/sgkim94/item05/PrintNotification.java
@@ -1,4 +1,4 @@
-package com.example.sgkim94.item04;
+package com.example.sgkim94.item05;
 
 public class PrintNotification implements Notification {
     public static final String PRINT_NOTIFY_MESSAGE = "Check Print!";
