+++ sgkim94/src/main/java/com/example/sgkim94/item05/SecurityNotification.java
@@ -1,4 +1,4 @@
-package com.example.sgkim94.item04;
+package com.example.sgkim94.item05;
 
 public class SecurityNotification implements Notification {
     public static final String SECURITY_NOTIFY_MESSAGE = "Check Security!";
