+++ sgkim94/src/main/java/com/example/sgkim94/item05/TimeoutNotification.java
@@ -1,4 +1,4 @@
-package com.example.sgkim94.item04;
+package com.example.sgkim94.item05;
 
 public class TimeoutNotification implements Notification {
     public static final String TIMEOUT_NOTIFY_MESSAGE = "Check Time!";
